agent_engine_agents/iap_classic_joke_agent_debug/pyproject.toml
[project]
name = "classic_joke_agent"
version = "0.1.0"
dependencies = [
    "google-adk",
    "requests",
    "google-auth"
]